feat(cache): Move code into __init__.py to simplify imports
Relocated functionality from a separate file into __init__.py to avoid extra path components when importing, making usage cleaner and
straightforward.
